solved 3rd problem lucky minimum
learned not to intialise array like this
int i ,a[i];
learned how to call array in a function
function(a,n);
/*my beta code of this problem and best submission in c was same */
editorials----------
Solution 1:
First find the minimum element in the array
A
A. The pseudo code is given below.

for i = 1 to N:
mi = min(mi, A[i])
Now calculate the frequency of the minimum element.
for i = 1 to N:
if A[i] == mi:
cnt = cnt + 1
Time Complexity:
O
(
N
)
O(N)
Solution 2:
We can combine the 2 loops in the previous solution and we can find the
answer in a single loop.

for i = 1 to N:
if mi > A[i]:
mi = A[i]
cnt = 1
else if m1 == a:
cnt = cnt + 1
Make sure to initialize
m
i
mi.
See the tester's code for more clarification.

Time Complexity:
O
(
N
)
O(N)
Solution 3:
This solution uses hashing. Count the frequencies of all the element in
a map and you will get the frequency of the minimum element as map
stores the element in sorted order. So the first element in the map is
the minimum element. The pseudo code below is similar to C++ code. See
the author's code for clarification.

map<int, int> m
for i = 1 to N:
m[A[i]]++
cnt = (*m.begin()).second;
This code will run in
O
(
N
∗
L
o
g
N
)
O(N∗LogN).