monk and search (first binary search problem took alot of time to solve )
Monk and Search

Monk and his friend Micro are on a quest to find the answer of Life,
Universe and Everything. In order to complete it they need to answer
Q
Q queries on an array
A
A having
N
N integers. The queries can be of following two types:
0

x
:
0 x: Find the number of numbers in
A
A which are not less than
x
x.
1

x
:
1 x: Find the number of numbers in
A
A which are greater than
x
x.
Help them complete the quest and be back in time for the next Code Monk
Challenge.

Input Format:
First line consists of a single integer denoting
N
N.
Second line consists of
N
N space separated integers denoting the array
A
A.
Third line consists of a single integer denoting
Q
Q.
Each of the following
Q
Q lines consists of a query of one of the given two types.

Output Format:
For each query print the required answer in new line.

Constraints:
1
≤
N
≤
10
5
1≤N≤105
1
≤
Q
≤
3
×
10
5
1≤Q≤3×105
1
≤
A
[
i
]
,
x
≤
10
9
1≤A[i],x≤109

SAMPLE INPUT
4
1 2 3 4
3
0 5
1 3
0 3
SAMPLE OUTPUT
0
1
2
Explanation
For first query, there are no numbers in the array which are not less
than
5
5 so answer for first query is
0
0.
For second query,
4
4 is the only number greater than
3
3, so answer is
1
1.
For third query,
3
3 and
4
4 are the numbers in array which are not less than
3
3, so answer is
2
2.
one test case didnt pass on hacker earth .but it was passing on ide one.
might be becs i ws using upper bound and lower bound